Added Linux version
How to build the Linux version and an executable that was built and tested in Ubuntu 20.04.5.  I was unable to verify how good it works as I do not have any AMD cards in that system and the Nvidia did not run fast enough to see if it was working.  Please let me know if it works ok or has a problem.